# Backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TaskFlow API",
    "version": "1.0.0",
    "description": "Complete REST API for TaskFlow - Auth, Workspaces, Tasks, Comments, and Attachments.",
    "contact": {
      "name": "TaskFlow Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "http://172.16.29.39:3000",
      "description": "Network development server"
    }
  ],
...
  "paths": {
...
    "/api/workspaces": {
      "post": {
        "tags": ["Workspaces"],
        "summary": "Create workspace",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              },
              "example": {
                "name": "Design Team",
                "description": "Workspace for product design"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workspace created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Workspaces"],
        "summary": "Get all workspaces for logged-in user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Workspaces fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspaces/{workspaceId}/tasks": {
      "post": {
        "tags": ["Tasks"],
        "summary": "Create task",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/WorkspaceIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              },
              "example": {
                "title": "Design login screen",
                "description": "Create high-fidelity mockups",
                "status": "TODO",
                "priority": "HIGH",
                "dueDate": "2026-05-01T12:00:00.000Z",
                "assigneeIds": ["ckz7x4u3p0000user001"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a workspace member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Tasks"],
        "summary": "Get all tasks in workspace",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/WorkspaceIdParam"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TaskStatus"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TaskPriority"
            }
          },
          {
            "name": "assigneeId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["dueDate", "priority", "createdAt"],
              "default": "createdAt"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Not a workspace member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspaces/{workspaceId}/tasks/{taskId}/comments": {
      "post": {
        "tags": ["Comments"],
        "summary": "Create comment on task",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/WorkspaceIdParam"},
          {"$ref": "#/components/parameters/TaskIdParam"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CreateCommentRequest"},
              "example": {"content": "Great work on this!"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created successfully",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommentResponse"}}}
          },
          "401": {
            "description": "Unauthorized",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "403": {
            "description": "Not a workspace member",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "404": {
            "description": "Task not found",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "422": {
            "description": "Validation failed",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          }
        }
      },
      "get": {
        "tags": ["Comments"],
        "summary": "Get all comments on task",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/WorkspaceIdParam"},
          {"$ref": "#/components/parameters/TaskIdParam"}
        ],
        "responses": {
          "200": {
            "description": "Comments fetched successfully",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommentListResponse"}}}
          },
          "401": {
            "description": "Unauthorized",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "403": {
            "description": "Not a workspace member",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "404": {
            "description": "Task not found",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          }
        }
      }
    },
...
    "/api/workspaces/{workspaceId}/tasks/{taskId}/attachments": {
      "post": {
        "tags": ["Attachments"],
        "summary": "Upload file attachment to task",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/WorkspaceIdParam"},
          {"$ref": "#/components/parameters/TaskIdParam"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {"type": "string", "format": "binary", "description": "File to upload"}
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded successfully",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AttachmentResponse"}}}
          },
          "400": {
            "description": "No file uploaded",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "401": {
            "description": "Unauthorized",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "403": {
            "description": "Not a workspace member",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "404": {
            "description": "Task not found",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          }
        }
      },
      "get": {
        "tags": ["Attachments"],
        "summary": "Get all attachments on task",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/WorkspaceIdParam"},
          {"$ref": "#/components/parameters/TaskIdParam"}
        ],
        "responses": {
          "200": {
            "description": "Attachments fetched successfully",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AttachmentListResponse"}}}
          },
          "401": {
            "description": "Unauthorized",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "403": {
            "description": "Not a workspace member",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          },
          "404": {
            "description": "Task not found",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use access token from /api/auth/login. Example: Bearer <token>"
      }
    },
    "parameters": {
      "WorkspaceIdParam": {
        "name": "workspaceId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Workspace ID"
      },
...
      "TaskIdParam": {
        "name": "taskId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Task ID"
      },
...
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" }
        },
        "required": ["success", "message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "errors": {
            "oneOf": [
              { "type": "object" },
              { "type": "array", "items": {} },
              { "type": "string" }
            ]
          }
        },
        "required": ["success", "message"]
      },
...
      "Workspace": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "ckz7x4u3p0000abc12345" },
          "name": { "type": "string", "example": "Design Team" },
          "description": { "type": "string", "nullable": true, "example": "Workspace for product design" },
          "ownerId": { "type": "string", "example": "ckz7x4u3p0000owner123" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true }
        },
        "required": ["id", "name", "ownerId", "createdAt", "updatedAt"]
      },
      "WorkspaceSummary": {
        "allOf": [
          { "$ref": "#/components/schemas/Workspace" },
          {
            "type": "object",
            "properties": {
              "role": {
                "type": "string",
                "enum": ["OWNER", "MEMBER"],
                "example": "OWNER"
              },
              "joinedAt": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": ["role", "joinedAt"]
          }
        ]
      },
...
      "WorkspaceResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/Workspace"
              }
            },
            "required": ["data"]
          }
        ]
      },
      "WorkspaceListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/WorkspaceSummary"
                }
              }
            },
            "required": ["data"]
          }
        ]
      },
...
      "CreateWorkspaceRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": ["name"]
      },
...
      "TaskStatus": {
        "type": "string",
        "enum": ["TODO", "IN_PROGRESS", "IN_REVIEW", "DONE"]
      },
      "TaskPriority": {
        "type": "string",
        "enum": ["LOW", "MEDIUM", "HIGH", "URGENT"]
      },
      "TaskUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "ckz7x4u3p0000user001" },
          "name": { "type": "string", "example": "Jane Member" },
          "email": { "type": "string", "format": "email", "example": "member@example.com" }
        },
        "required": ["id", "name", "email"]
      },
      "TaskAssignee": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "ckz7x4u3p0000assign001" },
          "taskId": { "type": "string", "example": "ckz7x4u3p0000task001" },
          "userId": { "type": "string", "example": "ckz7x4u3p0000user001" },
          "assignedAt": { "type": "string", "format": "date-time" },
          "user": { "$ref": "#/components/schemas/TaskUser" }
        },
        "required": ["id", "taskId", "userId", "assignedAt", "user"]
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "ckz7x4u3p0000task001" },
          "title": { "type": "string", "example": "Design login screen" },
          "description": { "type": "string", "nullable": true, "example": "Create high-fidelity mockups" },
          "status": { "$ref": "#/components/schemas/TaskStatus" },
          "priority": { "$ref": "#/components/schemas/TaskPriority" },
          "dueDate": { "type": "string", "format": "date-time", "nullable": true },
          "workspaceId": { "type": "string", "example": "ckz7x4u3p0000ws001" },
          "createdById": { "type": "string", "example": "ckz7x4u3p0000owner001" },
          "deletedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "createdBy": { "$ref": "#/components/schemas/TaskUser" },
          "assignees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskAssignee"
            }
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "priority",
          "workspaceId",
          "createdById",
          "createdAt",
          "updatedAt",
          "createdBy",
          "assignees"
        ]
      },
      "TaskPagination": {
        "type": "object",
        "properties": {
          "hasNextPage": { "type": "boolean", "example": true },
          "nextCursor": { "type": "string", "nullable": true, "example": "ckz7x4u3p0000task010" },
          "limit": { "type": "integer", "example": 20 }
        },
        "required": ["hasNextPage", "nextCursor", "limit"]
      },
      "TaskListData": {
        "type": "object",
        "properties": {
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/TaskPagination"
          }
        },
        "required": ["tasks", "pagination"]
      },
      "TaskResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/Task"
              }
            },
            "required": ["data"]
          }
        ]
      },
      "TaskListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/TaskListData"
              }
            },
            "required": ["data"]
          }
        ]
      },
...
      "CreateTaskRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "maxLength": 5000
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/TaskPriority"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "assigneeIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "default": []
          }
        },
        "required": ["title"]
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "ckz7x4u3p0000comment001" },
          "content": { "type": "string", "example": "This looks great!" },
          "taskId": { "type": "string", "example": "ckz7x4u3p0000task001" },
          "createdById": { "type": "string", "example": "ckz7x4u3p0000user001" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "createdBy": { "$ref": "#/components/schemas/TaskUser" }
        },
        "required": ["id", "content", "taskId", "createdById", "createdAt", "updatedAt", "createdBy"]
      },
      "CommentResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/Comment"
              }
            },
            "required": ["data"]
          }
        ]
      },
      "CommentListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            },
            "required": ["data"]
          }
        ]
      },
      "CreateCommentRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": ["content"]
      },
...
      "Attachment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "ckz7x4u3p0000attach001" },
          "originalName": { "type": "string", "example": "document.pdf" },
          "mimeType": { "type": "string", "example": "application/pdf" },
          "fileSize": { "type": "integer", "example": 1024000 },
          "taskId": { "type": "string", "example": "ckz7x4u3p0000task001" },
          "uploadedById": { "type": "string", "example": "ckz7x4u3p0000user001" },
          "createdAt": { "type": "string", "format": "date-time" },
          "uploadedBy": { "$ref": "#/components/schemas/TaskUser" }
        },
        "required": ["id", "originalName", "mimeType", "fileSize", "taskId", "uploadedById", "createdAt", "uploadedBy"]
      },
      "AttachmentResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/Attachment"
              }
            },
            "required": ["data"]
          }
        ]
      },
      "AttachmentListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessResponse" },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Attachment"
                }
              }
            },
            "required": ["data"]
          }
        ]
      }
    }
  }
}